+++ react/src/js/components/Consonant/Cards/OneHalf.jsx
@@ -14,10 +14,7 @@ import prettyFormatDate from '../Helpers/prettyFormat';
 import { INFOBIT_TYPE } from '../Helpers/constants';
 import { hasTag } from '../Helpers/Helpers';
 import { getEventBanner, getLinkTarget } from '../Helpers/general';
-import {
-    useConfig,
-    useLazyLoading,
-} from '../Helpers/hooks';
+import { useConfig } from '../Helpers/hooks';
 import {
     stylesType,
     contentAreaType,
@@ -175,23 +172,6 @@ const OneHalfCard = (props) => {
     });
 
     /**
-     * Creates a card image DOM reference
-     * @returns {Object} - card image DOM reference
-     */
-    const imageRef = React.useRef();
-
-    /**
-     * @typedef {Image} LazyLoadedImageState
-     * @description — Has image as state after image is lazy loaded
-     *
-     * @typedef {Function} LazyLoadedImageStateSetter
-     * @description - Sets state once image is lazy loaded
-     *
-     * @type {[Image]} lazyLoadedImage
-     */
-    const [lazyLoadedImage] = useLazyLoading(imageRef, image);
-
-    /**
      * Formatted date string
      * @type {String}
      */
@@ -282,8 +262,7 @@ const OneHalfCard = (props) => {
             <div
                 data-testid="consonant-OneHalfCard-img"
                 className="consonant-OneHalfCard-img"
-                ref={imageRef}
-                style={{ backgroundImage: lazyLoadedImage && `url("${lazyLoadedImage}")` }}
+                style={{ backgroundImage: `url("${image}")` }}
                 role={altText && 'img'}
                 aria-label={altText}>
                 {hasBanner && !disableBanners &&