+++ react/src/js/components/Consonant/Pagination/Paginator.jsx
@@ -10,6 +10,7 @@ import {
     generateRange,
     getStartNumber,
     getPageStartEnd,
+    getGlobalNavHeight,
 } from '../Helpers/general';
 
 const paginatorType = {
@@ -84,6 +85,11 @@ const Paginator = (props) => {
     const nextPageNotOutOfBounds = currentPageNumber + 1 < totalPages;
 
     /**
+     * GlobalNab height needed for scrolling
+     */
+    const globalNavHeight = getGlobalNavHeight();
+
+    /**
      * Handles click of prev, next or number button
      *
      * @param {ClickEvent} e
@@ -105,9 +111,11 @@ const Paginator = (props) => {
         } else {
             nextPage = parseInt(target.firstChild.nodeValue, BASE_10);
         }
-        const caasWrapper = target.closest('.consonant-Wrapper');
-        if (caasWrapper && typeof caasWrapper.scrollIntoView === 'function') {
-            caasWrapper.scrollIntoView({ behavior: 'smooth' });
+        const caasWrapper = target.closest('.section') || target.closest('.consonant-Wrapper');
+        if (caasWrapper
+            && caasWrapper.getBoundingClientRect().y < 0
+            && typeof caasWrapper.scrollIntoView === 'function') {
+            window.scrollTo({ left: 0, top: caasWrapper.offsetTop - globalNavHeight, behavior: 'smooth' });
         }
         onClick(nextPage);
     };